GEODE-2206: Add junit-quickcheck to geode-core.
* Rewrite a data serialization test to use junit-quickcheck.
